requirements.txt
chevron
pillow==11.2.1
pixelmatch==0.3.0
